Update LICENSE
BilikoyX Shell Spawner End User License Agreement (EULA)
BilikoyX grants you the right to download, use, and distribute the BilikoyX Shell Spawner (hereafter referred to as “Project”, “Code”, “Software”), provided the following terms and conditions are met:

Attribution: You agree to give appropriate credit to the original author. This includes mentioning the author in any documentation or promotional materials and providing a link to the project repository.

No Renaming: You are not permitted to rename, rebrand, or alter the identity of the Project in any way that may confuse users or misrepresent the origin of the software.

No Derivatives: You may not create any product, service, or derivative work based on any part of the Code, whether for commercial gain or free distribution. This ensures the integrity and originality of the Project is maintained.

No Re-licensing: You agree not to re-license the Code under any other terms or conditions. The original licensing terms must be maintained to protect the rights of the author.

Legal Use Only: You agree to use the Code only for lawful purposes. You must ensure that your use of the software complies with all applicable laws in your jurisdiction, as well as any laws where the software is being used or targets.

No Warranty: This Software is provided "as-is" without any warranty of any kind, express or implied. The author does not guarantee that the Software will be error-free, uninterrupted, or meet your specific requirements. The author takes no responsibility for any damages arising from the use or inability to use this Product, including but not limited to loss of data, profits, or other intangible losses.

Ownership: The Software code, name, and logos are owned by the original author and are protected by copyright and intellectual property laws. All rights not expressly granted in this EULA are reserved by the author.

Modification of Terms: The author reserves the right to modify the licensing terms at any time, without prior notice. It is your responsibility to periodically review the terms to remain compliant.

Severability: If any provision of this EULA is found to be unlawful, void, or unenforceable, the remaining provisions shall still be valid and enforceable. The invalid provision will be deemed severed from this EULA, and the validity of the other provisions will remain unaffected.

Acceptance: By downloading, using, or distributing this Project/Code, you acknowledge that you have read, understood, and agree to comply with this EULA. If you do not agree with any part of this agreement, you must cease use of the Project immediately and remove all copies from your systems.

For any questions concerning this EULA, please submit an issue in the project repository or contact the author directly.